Major clean-up and algorithmic changes.
1. Each converge request and notification runs in its own thread.
2. The stack lock is acquired after a eventlet thread is launched, hence
not blocking the RPC initiated thread.
3. Request ID is generated after every new request to update/delete a
stack. This helps workers ignore job requests for older stacks and helps
engine to drop the notification from worker when the job is done. The
worker and engine look for the request ID and decide on whether to
proceed or not.
4. Resources are created when they are needed to be converged. This is
change from previous approach where all the resources were created at
the beginning itself.
5. Added scheduled state to resource to identify resource request in
transit to workers.
6. Delete always happens after updating resources, even if rollback is
not enabled.
7. Purge all the edges not in the current graph in the end instead of
deleting them as and when the resources are deleted.